tk: stop colour values sign-extending to 64-bit
Every Tk colour is built as (R<<24)|(G<<16)|(B<<8)|A with R an int.
When the red byte has its top bit set (R >= 0x80, e.g. #CCCCCC, #E8553A,
white) the shift makes a negative int, and because WORD is pointer-sized
here it sign-extends on assignment to the ulong colour: 0xCCCCCCFF became
0xFFFFFFFFCCCCCCFF. Rendering survived (the draw layer uses the low 32
bits) but it broke anything that inspects the whole value:

  - `widget cget -background` returned "#ffffffffccccccff" instead of
    "#ccccccff", so apps round-tripping a colour through cget got an
    unparseable 16-digit string;
  - tkcolor()'s DWhite/DBlack fast paths never matched (white was
    0xFFFFFFFFFFFFFFFF, not 0xFFFFFFFF), so even pure white/black
    allocated and cached a 1x1 image.

Cast the top channel to ulong at each build site — tkrgba(),
tkparsecolor()'s #RGB/#RRGGBB/#RRGGBBAA cases, and colrs.c's RGB() macro
— so colours stay clean 32-bit (0x00000000rrggbbaa).

Verified: cget now returns "#ccccccff" / "#12abcdff" / "#080808ff" for
default and explicit colours.

+++ libtk/colrs.c
@@ -2,7 +2,7 @@
 #include "draw.h"
 #include "tk.h"
 
-#define RGB(R,G,B) (((R)<<24)|((G)<<16)|((B)<<8)|(0xff))
+#define RGB(R,G,B) (((ulong)(R)<<24)|((G)<<16)|((B)<<8)|(0xff))
 
 /*
  * Default Tk colour palette.
